Initial example for FullCalendar v1.4.6
http://arshaw.com/fullcalendar/

The example will draw a calendar, allow you to create events, store those events in your image, change the length of the events and start time, as well as delete the events.  

I plan to update this when FullCalendar 1.5 is released.  

Most features work.

You specify events by sending collectionOfEvents: a collection of JQFullCalendarEvent events (or a subclass you create).  The plugin also has the ability to get events from a JSON Feed or from a callback function to the server, but I haven't implemented anything to do that yet.