changed the reference in the associated style sheet to the stars gif from jquery.ui.stars.gif to: jqueryuistars.gif.
The file library handled the translation between jquery.ui.stars.gif and the Smalltalk method #jqueryuistarsGif without a problem, however if you are using a web server to deliver the file and you've exported the gif from the image, there will be a mismatch in names, resulting in a 404. By changing the reference in the css to match the exported name of the gif, all is when deploying the code.